requirements.txt
# Core dependencies
langchain>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0

# AI Provider dependencies
# Install the ones you want to use:

# Google Gemini (default)
langchain-google-genai>=2.1.0

# OpenAI (optional)
# langchain-openai>=0.3.0

# Anthropic Claude (optional)  
# langchain-anthropic>=0.3.0

# Enhanced features (optional)
# Uncomment the ones you want to use:

# Web interface
# streamlit>=1.28.0

# Voice capabilities
# speech-recognition>=3.10.0
# pyttsx3>=2.90

# Document processing (RAG)
# pypdf>=3.0.0
# faiss-cpu>=1.7.0
# langchain-community>=0.3.0

# Database support
# sqlalchemy>=2.0.0

# Configuration management
# pyyaml>=6.0

# Image processing
# pillow>=10.0.0

# Web scraping and search
# beautifulsoup4>=4.12.0
# requests>=2.31.0

# Progress bars and utilities
# tqdm>=4.66.0

# Testing
# pytest>=7.4.0
# pytest-asyncio>=0.21.0